+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1 - 16/08/2022
+## 0.0.2
+
+- Add `vimeo` feature to get video links via vimeo id.
+- Add `youtube` feature to get YouTube video links via youtube id.
+
+## 0.0.1
 
 - First patch version release.